Cut down managers into separate files
